combine test file config
