--- test/frontend/support/_bootstrap.js
@@ -1,4 +1,0 @@
-global.Notification = require('./_mockNotification');
-global.document = require('./_mockDocument');
-
-require('../setup')
